Refactor Fluent Bit deletion
Move Fluent Bit daemon set deletion code to logpipeline package since it
is nowhere else used.
